# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Documentação da API do W-AZAP",
    "version": "1.6.4",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor da API"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API gerada no painel (Webhooks e API)"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Cookie de login do painel (somente navegador)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error occurred"
          },
          "error": {
            "type": "string",
            "example": "Detailed error info"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autenticado: chave de API ausente ou inválida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Proibido: acesso negado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Forbidden - Cannot access this session",
              "error": "Forbidden - Cannot access this session"
            }
          }
        }
      },
...
      "SessionNotReady": {
        "description": "Sessão desconectada ou ainda não pronta",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": false,
              "message": "Session not ready",
              "error": "Session not ready"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/messages/{sessionId}/{jid}/media": {
      "post": {
        "tags": [
          "Mensagens"
        ],
        "summary": "Enviar mídia (imagem/vídeo/áudio/documento)",
        "description": "Envia um arquivo via `multipart/form-data` (campos `file`, `type` e `caption`). Tamanho máximo: `MAX_UPLOAD_SIZE_MB`.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identificador da sessão"
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "JID do destinatário"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "type"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video",
                      "audio",
                      "voice",
                      "document",
                      "sticker"
                    ],
                    "default": "image"
                  },
                  "caption": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mídia enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida: arquivo ausente"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Falha ao enviar a mídia"
          },
          "503": {
            "$ref": "#/components/responses/SessionNotReady"
          }
        }
      }
    },
...
    "/messages/{sessionId}/{jid}/{messageId}/reply": {
      "post": {
        "tags": [
          "Mensagens"
        ],
        "summary": "Responder a uma mensagem (com citação)",
        "description": "Envia uma resposta citando uma mensagem específica pelo ID. Usa o mesmo formato do `/send`: envie um objeto de mensagem do Baileys.",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "sales-01"
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "5511987654321@s.whatsapp.net"
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID da mensagem a responder",
            "example": "3EB0ABCD1234567890"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "object",
                    "description": "Conteúdo da mensagem, no mesmo formato do `/send` (texto, imagem, vídeo etc.)",
                    "example": {
                      "text": "Thanks for your message!"
                    }
                  },
                  "mentions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": [
                      "5511987654321@s.whatsapp.net"
                    ],
                    "description": "JIDs a mencionar"
                  },
                  "fromMe": {
                    "type": "boolean",
                    "default": false,
                    "description": "Indica se a mensagem citada foi enviada por você"
                  }
                }
              },
              "examples": {
                "textReply": {
                  "summary": "Resposta em texto",
                  "value": {
                    "message": {
                      "text": "Got it, thanks!"
                    }
                  }
                },
                "imageReply": {
                  "summary": "Resposta com imagem",
                  "value": {
                    "message": {
                      "image": {
                        "url": "https://example.com/confirm.jpg"
                      },
                      "caption": "Confirmation image"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta enviada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Message sent successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`message` é obrigatório"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Falha ao enviar a resposta"
          },
          "503": {
            "$ref": "#/components/responses/SessionNotReady"
          }
        }
      }
    },
...
  },
...
}